+++ Reconstruction/src/Variables.cxx
@@ -21,7 +21,7 @@ namespace Reconstruction{
 	std::string					inputDir; // To be removed when the whole drawout code is updated
 
 	// Run variables
-	std::vector<int>			v_valint;
+	std::vector<float>	 		v_valint;
 	std::vector<std::string>	v_valstr;
 	std::vector<std::string>	v_scanspec;
 	std::vector<std::string> 	v_comments;
